(int)(int)(int)(int)(int)(int)(int)

+++ Assets/Scripts/UI Scripts/PetStatsUI.cs
@@ -91,8 +91,8 @@ public class PetStatsUI : MonoBehaviour //Not finished yet lmao
                 OnPetStatsShown?.Invoke();
                 petNameText.text = pet.name;
                 petTypeText.text = pet.petType.ToString();
-                petHungerText.text = pet.hunger + "/" + pet.maxHunger;
-                petThirstText.text = pet.thirst + "/" + pet.maxThirst;
+                petHungerText.text = (int)pet.hunger + "/" + (int)pet.maxHunger;
+                petThirstText.text = (int)pet.thirst + "/" + (int)pet.maxThirst;
                 petHungerSlider.value = pet.hunger / pet.maxHunger;
                 petThirstSlider.value = pet.thirst / pet.maxThirst;
                 petFriendshipText.text = pet.friendshipLevel.ToString();
@@ -105,9 +105,9 @@ public class PetStatsUI : MonoBehaviour //Not finished yet lmao
                 critterNameText.text = critter.name;
                 critterHomeIcon.gameObject.SetActive(!critter.IsCritterHomeless());
                 critterTypeText.text = critter.creatureData.name.ToString();
-                critterHealthText.text = critter.health + "/" + critter.maxHealth;
-                critterHungerText.text = critter.hunger + "/" + critter.maxHunger;
-                critterThirstText.text = critter.thirst + "/" + critter.maxThirst;
+                critterHealthText.text = (int)critter.health + "/" + (int)critter.maxHealth;
+                critterHungerText.text = (int)critter.hunger + "/" + (int)critter.maxHunger;
+                critterThirstText.text = (int)critter.thirst + "/" + (int)critter.maxThirst;
                 critterHealthSlider.value = critter.health / critter.maxHealth;
                 critterHungerSlider.value = critter.hunger / critter.maxHunger;
                 critterThirstSlider.value = critter.thirst / critter.maxThirst;
